Make sure users are properly notified of (useless) exceptions in finally blocks
It is possible to rewrite Try::Tiny to invoke the finally blocks without
the use of scope guards, however to do this an extra eval frame around
catch will have to be introduced. Furthermore code in the wild may rely
on the "ignorance" of a finally() exception.

In short - the proper fix may no longer be feasible due to backcompat
issues, therefore the current warning seems like the only feasible
compromise.